+++ pkg/scheduler/cache/candidate/base.go
@@ -33,8 +33,6 @@ import (
 	schedmodels "yunion.io/x/onecloud/pkg/scheduler/models"
 )
 
-var ErrInstanceGroupNotFound = errors.Error("InstanceGroupNotFound")
-
 type BaseHostDesc struct {
 	*computemodels.SHost
 	Region        *computemodels.SCloudregion              `json:"region"`
@@ -138,7 +136,7 @@ func (b baseHostGetter) GetFreeGroupCount(groupId string) (int, error) {
 	// Must Be
 	scg, ok := b.h.InstanceGroups[groupId]
 	if !ok {
-		return 0, errors.Wrap(ErrInstanceGroupNotFound, groupId)
+		return 0, errors.Wrap(core.ErrInstanceGroupNotFound, groupId)
 	}
 	free := scg.Granularity - scg.ReferCount
 	if free < 1 {
